Move crop rename object to `/constants/crops`
